Frontend folder structure refactor
- moved all relevant components and pages into their respective folders under the pages folder
- global components stay in src/components
- Example: HeroSection.js is moved to src/pages/Home, along with Home.js
