+++ tecttitans-app/src/components/Footer.js
