+++ tecttitans-app/src/pages/Home/HeroSection.js
@@ -1,7 +1,7 @@
 import React from 'react';
-import '../App.css'
+import '../../App.css'
 import './HeroSection.css';
-import {Button} from "./Button";
+import {Button} from "../../components/Button";
 import hero from './HeroSection.css'
 import {Link} from 'react-router-dom';
