+++ tecttitans-app/src/pages/Home/Home.js
@@ -1,6 +1,8 @@
 import React from 'react';
-import '../App.css';
-import HeroSection from '../components/HeroSection';
+import '../../App.css';
+import HeroSection from './HeroSection';
+
+
 function Home() {
     return (
         <>
