+++ tecttitans-app/src/pages/Tour/PlaceDetails.css
