+++ tecttitans-app/src/pages/Tour/PlaceDetails.js
