+++ tecttitans-app/src/pages/Tour/Tour.js
@@ -1,13 +1,13 @@
 import React, {useState } from 'react';
-import '../App.css';
+import '../../App.css';
 
 import { CssBaseline } from '@mui/material';
 import { createTheme, ThemeProvider } from '@mui/material/styles';
 
-import Header from '../maps/Header';
-import Map from '../maps/Map';
-import PlaceDetails from '../maps/PlaceDetails';
-import places from "../data/hotels_data.json"
+import Header from './Header';
+import Map from './Map';
+import PlaceDetails from './PlaceDetails';
+import places from "../../data/hotels_data.json"
 
 function Tour() {
